refactor: divide common, model code from main (#4)
