+++ Backend/.env.example
@@ -1,2 +1,4 @@
 SQLALCHEMY_DATABASE_URI=mysql://your_username:your_password@localhost:your_db_port/your_database
 USER_TABLE_NAME=table_name
+BEFORE_TABLE_NAME=table_name
+AFTER_TABLE_NAME=table_name
