+++ Backend/requirements.txt
@@ -9,6 +9,7 @@ Jinja2==3.1.2
 MarkupSafe==2.1.3
 mysqlclient==2.2.0
 python-dotenv==1.0.0
+pytz==2023.3
 SQLAlchemy==2.0.18
 typing_extensions==4.7.1
 Werkzeug==2.3.6
